file_io

+++ 0x15-file_io/0-read_textfile.c
@@ -12,7 +12,7 @@
 
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	int i, ch;
+	size_t i, ch;
 	FILE *fptr;
 
 	fptr = fopen(filename, "r");
